Unpark the ddem.ca domain name
The domain is currently parked. This index.html file will replace what's there now.

The intention is to upgrade this to a Terraform setup in coming changes. The reason for Terraform is to break the founder's veto. The entirety of ddem.ca including the details of which servers to provision and how, will be under full democratic control. As we experiment with Terraform, we continue to look at Ethereum DAO and other technology for improved solutions to break the founder's veto.